pyproject.toml
[project]
name = "caisseaoutils"
version = "0.1.0"
description = "Community toolbox of Django resources for Django Cameroon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "markdown>=3.9",
    "pyyaml>=6.0.2",
]
